# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bp-carnet",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/payments/apple/account-token": {
      "post": {
        "operationId": "apple_account_token",
        "responses": {
          "200": {
            "description": "アカウントの appAccountToken (無ければ作った)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppleAccountTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "アプリ内課金が設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/apple/notifications": {
      "post": {
        "summary": "App Store Server Notifications V2。署名は確かめず、取引を Apple から取り直して反映する。\n認証・CSRF の対象にしない理由は `stripe_webhook` と同じ。",
        "operationId": "apple_notification",
        "requestBody": {
          "description": "responseBodyV2 (`signedPayload`)",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "反映した、または関係の無い通知"
          },
          "400": {
            "description": "形が不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "送り元が Apple の範囲ではない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "アプリ内課金が設定されていない、または今は確かめられない (再送させる)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/apple/transactions": {
      "post": {
        "summary": "枠は取引の `appAccountToken` の持ち主に付く (送ってきたアカウントとは限らない)。",
        "operationId": "apple_transaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "取引を Apple から取り直して反映した",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppleTransactionResponse"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "送る回数の上限を超えた (アプリは取引を終えず、次の起動で送り直す)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "アプリ内課金が設定されていない (payments_disabled)、または今は確かめられない (app_store_unavailable)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/ocr-topup/checkout": {
      "post": {
        "operationId": "ocr_topup_checkout",
        "responses": {
          "200": {
            "description": "Checkout Session を作成した",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrTopupCheckoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Stripe API 呼び出しに失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe 決済が設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/ocr-topup/result": {
      "get": {
        "operationId": "ocr_topup_result",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "description": "Checkout Session ID (成功URLの `{CHECKOUT_SESSION_ID}` から)。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "購入結果 (成功・処理待ち・失敗)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrTopupResultResponse"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "本人の Session ではない、または存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Stripe API 呼び出しに失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe 決済が設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/stripe/webhook": {
      "post": {
        "summary": "Stripe からの Webhook。署名 (`Stripe-Signature`) 検証は生のリクエストボディに対して行うため、\nJSON 抽出 (`Json<T>`) ではなく `Bytes` で受ける (JSON を経由して再シリアライズすると\n署名が一致しなくなる)。\nブラウザ由来ではないため認証・CSRF 同一オリジンチェックの対象にしない\n(`crate::csrf::same_origin_check` は Origin/Referer が無いリクエストを素通りさせる設計)。",
        "operationId": "stripe_webhook",
        "requestBody": {
          "description": "Stripe Event (生の JSON。署名検証のため型付けしない)",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "受理・処理した (未知のイベント種別も含む)"
          },
          "400": {
            "description": "署名が不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Session の再検証等、Stripe API 呼び出しに失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe 決済が設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppStoreTransactionStatus": {
        "type": "string",
        "description": "アプリから届いた取引を、Apple から取り直して枠に反映した結果。",
        "enum": [
          "granted",
          "already_granted",
          "rejected"
        ]
      },
      "AppleAccountTokenResponse": {
        "type": "object",
        "required": [
          "appAccountToken"
        ],
        "properties": {
          "appAccountToken": {
            "type": "string",
            "description": "購入オプションの `appAccountToken` に渡す UUID (小文字)。アカウントごとに固定。"
          }
        }
      },
...
      "AppleTransactionRequest": {
        "type": "object",
        "required": [
          "transactionId",
          "environment"
        ],
        "properties": {
          "environment": {
            "type": "string",
            "description": "StoreKit の `Transaction.environment` の値 (`Production`・`Sandbox`・`Xcode`)。先にこの環境の API から引く。\n`Xcode` (StoreKit のローカルのテスト) の取引は Apple に無いので `rejected` になる。"
          },
          "transactionId": {
            "type": "string",
            "description": "StoreKit の `Transaction.id`。"
          }
        }
      },
      "AppleTransactionResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/AppStoreTransactionStatus",
            "description": "どれでもアプリは取引を終える (`finish()`)。エラーの応答なら終えない。"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "rows": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/RowError"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "レスポンスボディの共通 envelope。`code` が機械可読な契約 (frontend はこちらで表示文言を引く)、\n`message` はローカライズしないデバッグ用の説明文。`rows` は `BatchValidation` のときだけ埋まる。",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "OcrTopupCheckoutResponse": {
        "type": "object",
        "required": [
          "checkoutUrl"
        ],
        "properties": {
          "checkoutUrl": {
            "type": "string",
            "description": "この URL へ遷移すると Stripe-hosted Checkout が開く。"
          }
        }
      },
      "OcrTopupResultResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OcrTopupResultStatus"
          }
        }
      },
      "OcrTopupResultStatus": {
        "type": "string",
        "enum": [
          "processing",
          "succeeded",
          "failed"
        ]
      },
...
      "RowError": {
        "type": "object",
        "description": "バッチ処理で 1 行の検証が失敗したことを表す。`index` はリクエスト配列内の 0 始まりの位置。",
        "required": [
          "index",
          "message"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}